+++ src/auth/azureAd.ts
@@ -15,9 +15,30 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
+import axios from 'axios';
 import { Express, RequestHandler } from 'express';
 import { PassportStatic } from 'passport';
 import { IBearerStrategyOptionWithRequest } from 'passport-azure-ad';
+import { RequestError } from '../errors/requestError';
+
+/**
+ * An object of a current user in Microsoft Graph API.
+ *
+ * @see https://docs.microsoft.com/en-us/graph/api/user-get?view=graph-rest-1.0&tabs=http
+ */
+export interface IMicrosoftMe {
+    businessPhones?: string[];
+    displayName?: string;
+    givenName?: string;
+    id: string;
+    jobTitle?: string;
+    mail?: string;
+    mobilePhone?: string;
+    officeLocation?: string;
+    preferredLanguage?: string;
+    surname?: string;
+    userPrincipalName: string;
+}
 
 /**
  * Result of a 'setupForAzureAD' function call.
@@ -27,6 +48,31 @@ export interface ISetupForAzureADResult {
      * Creates a request handler, which requires a valie Azure AD Bearer token.
      */
     withAzureADBearer: () => RequestHandler;
+}
+
+/**
+ * Returns information about the current user, using an access token.
+ *
+ * @param {string} accessToken The access token.
+ *
+ * @returns {Promise<IMicrosoftMe>} The promise with the access token.
+ */
+export async function getMicrosoftMe(accessToken: string): Promise<IMicrosoftMe> {
+    try {
+        const response = await axios.get('https://graph.microsoft.com/v1.0/me', {
+            headers: {
+                Authorization: `Bearer ${accessToken}`
+            }
+        });
+
+        if (response.status !== 200) {
+            throw new RequestError(response, `Unexpected response ${response.status}`);
+        }
+
+        return response.data;
+    } catch (ex) {
+        throw new RequestError(ex, ex.message);
+    }
 }
 
 /**
@@ -80,4 +126,19 @@ export function setupAppForAzureAD(app: Express): ISetupForAzureADResult {
     return {
         withAzureADBearer: () => passport.authenticate('oauth-bearer', { session: false })
     };
+}
+
+/**
+ * Tries to return information about the current user, using an access token.
+ *
+ * @param {string} accessToken The access token.
+ *
+ * @returns {Promise<IMicrosoftMe | false>} The promise with the access token or (false) if not possible.
+ */
+export async function tryGetMicrosoftMe(accessToken: string): Promise<IMicrosoftMe | false> {
+    try {
+        return await getMicrosoftMe(accessToken);
+    } catch {
+        return false;
+    }
 }